+++ packages/lumx-vue/.eslintrc.cjs
@@ -1,0 +1,17 @@
+module.exports = {
+  extends: [
+    "../../.eslintrc.json"
+  ],
+  plugins: ["vue"],
+  parser: "vue-eslint-parser",
+  parserOptions: {
+    parser: "@typescript-eslint/parser",
+    project: "./tsconfig.json",
+    tsconfigRootDir: __dirname,
+    extraFileExtensions: [".vue"]
+  },
+  rules: {
+    "vue/multi-word-component-names": "off",
+    "@typescript-eslint/no-unused-vars": "off"
+  }
+};