+++ packages/lumx-vue/src/components/icon/index.ts
@@ -1,0 +1,8 @@
+import { IconProps, IconSizes, IconClassName } from '@lumx/core/js/components/Icon';
+
+import Icon from './Icon.vue';
+
+Icon.className = IconClassName;
+
+export { Icon };
+export type { IconProps, IconSizes };